secu user ok

+++ assets/actions/delete_notification_action.php
@@ -2,10 +2,24 @@
 session_start();
 require('../require/check_data.php');
 
-require('../require/co_bdd.php');
 
-$req = $bdd->prepare('DELETE FROM notifications WHERE id = :id');
-$req->bindParam(':id', $_GET['id'], PDO::PARAM_INT);
-$req->execute();
+if (
+  isset($get_id)
+  && isset($session_users_id)
+) {
 
-header('location: ../../my_account.php?success=notification_deleted');
+  require('../require/co_bdd.php');
+
+  $req = $bdd->prepare('DELETE FROM notifications WHERE id = :id');
+  $req->bindParam(':id', $get_id, PDO::PARAM_INT);
+  $req->execute();
+
+  $bdd = null;
+  header('location: ../../my_account.php?success=notification_deleted');
+} else {
+
+  $bdd = null;
+  http_response_code(400);
+  header('location: ../../my_account.php?error=processing_bad_or_malformed_request');
+  die();
+}
